Move AppState into its own module
